ENT-1314: Modify the locking and transaction size for the hypervisor report
- Split transaction into a series of bulk operations per 10 items
- Extract logic from job into HypervisorUpdateAction to split it into a separate transaction
