+++ server/src/main/java/org/candlepin/pinsetter/tasks/HypervisorUpdateJob.java
@@ -14,38 +14,31 @@
  */
 package org.candlepin.pinsetter.tasks;
 
-import static org.quartz.JobBuilder.*;
+import static org.quartz.JobBuilder.newJob;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.inject.name.Named;
 import org.candlepin.auth.Principal;
 import org.candlepin.common.exceptions.BadRequestException;
 import org.candlepin.common.filter.LoggingFilter;
 import org.candlepin.dto.ModelTranslator;
+import org.candlepin.dto.api.v1.ConsumerDTO;
 import org.candlepin.dto.api.v1.HypervisorConsumerDTO;
+import org.candlepin.dto.api.v1.HypervisorUpdateResultDTO;
 import org.candlepin.model.Consumer;
 import org.candlepin.model.ConsumerCurator;
-import org.candlepin.model.ConsumerType;
-import org.candlepin.model.ConsumerType.ConsumerTypeEnum;
-import org.candlepin.model.ConsumerTypeCurator;
-import org.candlepin.model.GuestId;
-import org.candlepin.model.HypervisorId;
 import org.candlepin.model.JobCurator;
 import org.candlepin.model.Owner;
 import org.candlepin.model.OwnerCurator;
 import org.candlepin.model.VirtConsumerMap;
 import org.candlepin.pinsetter.core.model.JobStatus;
-import org.candlepin.policy.js.compliance.ComplianceRules;
 import org.candlepin.resource.ConsumerResource;
-import org.candlepin.dto.api.v1.HypervisorUpdateResultDTO;
-import org.candlepin.resource.util.GuestMigration;
-import org.candlepin.service.SubscriptionServiceAdapter;
+import org.candlepin.service.impl.HypervisorUpdateAction;
 import org.candlepin.util.Util;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Lists;
 import com.google.inject.Inject;
-import com.google.inject.persist.Transactional;
+import com.google.inject.name.Named;
 
-import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.MDC;
 import org.quartz.JobDataMap;
 import org.quartz.JobDetail;
@@ -63,13 +56,11 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import java.util.zip.DeflaterOutputStream;
 import java.util.zip.InflaterInputStream;
@@ -83,36 +74,34 @@ public class HypervisorUpdateJob extends KingpinJob {
 
     private static Logger log = LoggerFactory.getLogger(HypervisorUpdateJob.class);
 
-    private ObjectMapper mapper;
     private OwnerCurator ownerCurator;
     private ConsumerCurator consumerCurator;
-    private ConsumerResource consumerResource;
-    private I18n i18n;
-    private ConsumerType hypervisorType;
-    private SubscriptionServiceAdapter subAdapter;
-    private ComplianceRules complianceRules;
     private ModelTranslator translator;
+    private HypervisorUpdateAction hypervisorUpdateAction;
+    private I18n i18n;
+    private ObjectMapper mapper;
 
     public static final String CREATE = "create";
-    public static final String REPORTER_ID = "reporter_id";
-    public static final String DATA = "data";
-    public static final String PRINCIPAL = "principal";
+    private static final String REPORTER_ID = "reporter_id";
+    private static final String DATA = "data";
+    private static final String PRINCIPAL = "principal";
     protected static String prefix = "hypervisor_update_";
+    private static final int BULK_SIZE = 10;
 
     @Inject
-    public HypervisorUpdateJob(OwnerCurator ownerCurator, ConsumerCurator consumerCurator,
-        ConsumerTypeCurator consumerTypeCurator, ConsumerResource consumerResource, I18n i18n,
-        SubscriptionServiceAdapter subAdapter, ComplianceRules complianceRules, ModelTranslator translator,
-        @Named("HypervisorUpdateJobObjectMapper") ObjectMapper objectMapper) {
-        this.ownerCurator = ownerCurator;
-        this.consumerCurator = consumerCurator;
-        this.consumerResource = consumerResource;
-        this.i18n = i18n;
-        this.subAdapter = subAdapter;
-        this.complianceRules = complianceRules;
-        this.translator = translator;
-        this.hypervisorType = consumerTypeCurator.getByLabel(ConsumerTypeEnum.HYPERVISOR.getLabel(), true);
-        this.mapper = objectMapper;
+    public HypervisorUpdateJob(
+        final OwnerCurator ownerCurator,
+        final ConsumerCurator consumerCurator,
+        final ModelTranslator translator,
+        final HypervisorUpdateAction hypervisorUpdateAction,
+        final I18n i18n,
+        @Named("HypervisorUpdateJobObjectMapper") final ObjectMapper objectMapper) {
+        this.ownerCurator = Objects.requireNonNull(ownerCurator);
+        this.consumerCurator = Objects.requireNonNull(consumerCurator);
+        this.translator = Objects.requireNonNull(translator);
+        this.hypervisorUpdateAction = Objects.requireNonNull(hypervisorUpdateAction);
+        this.i18n = Objects.requireNonNull(i18n);
+        this.mapper = Objects.requireNonNull(objectMapper);
     }
 
     public static JobStatus scheduleJob(JobCurator jobCurator, Scheduler scheduler, JobDetail detail,
@@ -127,8 +116,7 @@ public class HypervisorUpdateJob extends KingpinJob {
         }
 
         log.debug("Scheduling job without a trigger: {}", detail.getKey().getName());
-        JobStatus status = KingpinJob.scheduleJob(jobCurator, scheduler, detail, null);
-        return status;
+        return KingpinJob.scheduleJob(jobCurator, scheduler, detail, null);
     }
 
     public static boolean isSchedulable(JobCurator jobCurator, JobStatus status) {
@@ -137,94 +125,33 @@ public class HypervisorUpdateJob extends KingpinJob {
         return nextJob != null && nextJob.getId().equals(status.getId());
     }
 
-    private void parseHypervisorList(HypervisorList hypervisorList, Set<String> hosts,
-        Set<String> guests, Map<String, Consumer> incomingHosts) {
-        int emptyGuestIdCount = 0;
-        int emptyHypervisorIdCount = 0;
-
-        List<Consumer> l = hypervisorList.getHypervisors();
-        for (Iterator<Consumer> hypervisors = l.iterator(); hypervisors.hasNext();) {
-            Consumer hypervisor = hypervisors.next();
-
-            HypervisorId idWrapper = hypervisor.getHypervisorId();
-
-            if (idWrapper == null) {
-                continue;
-            }
-
-            String id = idWrapper.getHypervisorId();
-
-            if (id == null) {
-                continue;
-            }
-
-            if ("".equals(id)) {
-                hypervisors.remove();
-                emptyHypervisorIdCount++;
-                continue;
-            }
-
-            incomingHosts.put(id, hypervisor);
-            hosts.add(id);
-
-            List<GuestId> guestsIdList = hypervisor.getGuestIds();
-
-            if (guestsIdList == null || guestsIdList.isEmpty()) {
-                continue;
-            }
-
-            for (Iterator<GuestId> guestIds = guestsIdList.iterator(); guestIds.hasNext();) {
-                GuestId guestId = guestIds.next();
-                if (StringUtils.isEmpty(guestId.getGuestId())) {
-                    guestIds.remove();
-                    emptyGuestIdCount++;
-                }
-                else {
-                    guests.add(guestId.getGuestId());
-                }
-            }
-        }
-
-        if (emptyHypervisorIdCount > 0) {
-            log.debug("Ignoring {} hypervisors with empty hypervisor IDs", emptyHypervisorIdCount);
-        }
-
-        if (emptyGuestIdCount > 0) {
-            log.debug("Ignoring {} empty/null guestId(s)", emptyGuestIdCount);
-        }
-    }
-
     /**
      * {@inheritDoc}
      *
-     * Executes {@link ConsumerResource#create(org.candlepin.model.Consumer, org.candlepin.auth.Principal,
-     *  java.lang.String, java.lang.String, java.lang.String)}
-     * Executes (@link ConusmerResource#performConsumerUpdates(java.utl.String, org.candlepin.model.Consumer)}
+     * Executes {@link ConsumerResource#create(ConsumerDTO, Principal, String, String, String, boolean)}
+     * Executes {@link ConsumerResource#updateConsumer(String, ConsumerDTO, Principal)}
      * as a pinsetter job.
      *
      * @param context the job's execution context
      */
-    @Transactional
-    @SuppressWarnings({"checkstyle:indentation", "checkstyle:methodlength"})
-    public void toExecute(JobExecutionContext context) throws JobExecutionException {
+    @SuppressWarnings({ "checkstyle:indentation", "checkstyle:methodlength" })
+    public void toExecute(final JobExecutionContext context) throws JobExecutionException {
         try {
-            JobDataMap map = context.getMergedJobDataMap();
-            String ownerKey = map.getString(JobStatus.TARGET_ID);
-            Boolean create = map.getBoolean(CREATE);
-            Principal principal = (Principal) map.get(PRINCIPAL);
-            String jobReporterId = map.getString(REPORTER_ID);
+            final JobDataMap map = context.getMergedJobDataMap();
+            final String ownerKey = map.getString(JobStatus.TARGET_ID);
+            final Boolean create = map.getBoolean(CREATE);
+            final Principal principal = (Principal) map.get(PRINCIPAL);
+            final String jobReporterId = map.getString(REPORTER_ID);
 
-            HypervisorUpdateResultDTO result = new HypervisorUpdateResultDTO();
-
-            Owner owner = ownerCurator.getByKey(ownerKey);
+            final Owner owner = ownerCurator.getByKey(ownerKey);
             if (owner == null) {
                 context.setResult("Nothing to do. Owner does not exist");
-                log.warn("Hypervisor update attempted against non-existent org id \"{0}\"", ownerKey);
+                log.warn("Hypervisor update attempted against non-existent org id \"{}\"", ownerKey);
                 return;
             }
 
             if (owner.isAutobindDisabled() || owner.isContentAccessEnabled()) {
-                String caMessage = owner.isContentAccessEnabled() ?
+                final String caMessage = owner.isContentAccessEnabled() ?
                     " because of the content access mode setting" : "";
                 log.debug("Could not update host/guest mapping. Auto-Attach is disabled for owner {}{}",
                     owner.getKey(), caMessage);
@@ -233,129 +160,27 @@ public class HypervisorUpdateJob extends KingpinJob {
                         owner.getKey(), caMessage));
             }
 
-            byte[] data = (byte[]) map.get(DATA);
-            String json = decompress(data);
-            HypervisorList hypervisors = mapper.readValue(json, HypervisorList.class);
-            log.debug("Hypervisor consumers for create/update: {}", hypervisors.getHypervisors().size());
-            log.debug("Updating hypervisor consumers for org {}", ownerKey);
+            final HypervisorList hypervisors = parsedHypervisors(map);
+            final HypervisorUpdateAction.Result updateResult = hypervisorUpdateAction.update(
+                owner, hypervisors, create, principal, jobReporterId);
+            final HypervisorUpdateResultDTO result = updateResult.getResult();
+            final VirtConsumerMap hypervisorKnownConsumersMap = updateResult.getKnownConsumers();
 
-            Set<String> hosts = new HashSet<>();
-            Set<String> guests = new HashSet<>();
-            Map<String, Consumer> incomingHosts = new HashMap<>();
-            parseHypervisorList(hypervisors, hosts, guests, incomingHosts);
-            // TODO Need to ensure that we retrieve existing guestIds from the DB before continuing.
-
-            // Maps virt hypervisor ID to registered consumer for that hypervisor, should one exist:
-            VirtConsumerMap hypervisorKnownConsumersMap =
-                consumerCurator.getHostConsumersMap(owner, hypervisors);
-            Map<String, Consumer> systemUuidKnownConsumersMap = new HashMap<>();
+            final List<Consumer> created = new ArrayList<>();
+            final List<Consumer> updated = new ArrayList<>();
             for (Consumer consumer : hypervisorKnownConsumersMap.getConsumers()) {
-                if (consumer.hasFact(Consumer.Facts.SYSTEM_UUID)) {
-                    systemUuidKnownConsumersMap.put(
-                        consumer.getFact(Consumer.Facts.SYSTEM_UUID), consumer);
-                }
-            }
-
-            Map<String, GuestId> guestIds = consumerCurator.getGuestIdMap(guests, owner);
-            for (String hypervisorId : hosts) {
-                Consumer incoming = incomingHosts.get(hypervisorId);
-                Consumer knownHost = hypervisorKnownConsumersMap.get(hypervisorId);
-                // HypervisorId might be different in candlepin
-                if (knownHost == null && incoming.hasFact(Consumer.Facts.SYSTEM_UUID) &&
-                    systemUuidKnownConsumersMap.get(incoming.getFact(Consumer.Facts.SYSTEM_UUID)) != null) {
-                    knownHost = systemUuidKnownConsumersMap.get(incoming.getFact(Consumer.Facts.SYSTEM_UUID));
-                    if (knownHost != null) {
-                        log.debug("Found a known host by system uuid");
-                    }
-                }
-
-                Consumer reportedOnConsumer = null;
-
-                if (knownHost == null) {
-                    if (!create) {
-                        result.addFailed(hypervisorId,
-                            "Unable to find hypervisor with id " + hypervisorId + " in org " + ownerKey);
-                    }
-                    else {
-                        log.debug("Registering new host consumer for hypervisor ID: {}", hypervisorId);
-                        Consumer newHost = createConsumerForHypervisorId(hypervisorId, jobReporterId, owner,
-                            principal, incoming);
-
-                        // Since we just created this new consumer, we can migrate the guests immediately
-                        GuestMigration guestMigration = new GuestMigration(consumerCurator)
-                            .buildMigrationManifest(incoming, newHost);
-
-                        // Now that we have the new consumer persisted, immediately migrate the guests to it
-                        if (guestMigration.isMigrationPending()) {
-                            guestMigration.migrate(false);
-                        }
-
-                        hypervisorKnownConsumersMap.add(hypervisorId, newHost);
-                        result.addCreated(this.translator.translate(newHost, HypervisorConsumerDTO.class));
-                        reportedOnConsumer = newHost;
-                    }
+                final HypervisorConsumerDTO translated = this.translator.translate(
+                    consumer, HypervisorConsumerDTO.class);
+                if (result.wasCreated(translated)) {
+                    created.add(consumer);
                 }
                 else {
-                    boolean hypervisorIdUpdated = false;
-                    if (knownHost.getHypervisorId() != null && !hypervisorId.equalsIgnoreCase(knownHost
-                        .getHypervisorId().getHypervisorId())) {
-                        hypervisorIdUpdated = true;
-                        log.debug("Changing hypervisor id to [" + hypervisorId + "]");
-                        knownHost.getHypervisorId().setHypervisorId(hypervisorId);
-                    }
-
-                    reportedOnConsumer = knownHost;
-                    if (jobReporterId != null && knownHost.getHypervisorId() != null &&
-                        hypervisorId.equalsIgnoreCase(knownHost.getHypervisorId().getHypervisorId()) &&
-                        knownHost.getHypervisorId().getReporterId() != null &&
-                        !jobReporterId.equalsIgnoreCase(knownHost.getHypervisorId().getReporterId())) {
-                        log.debug("Reporter changed for Hypervisor {} of Owner {} from {} to {}",
-                            hypervisorId, ownerKey, knownHost.getHypervisorId().getReporterId(),
-                            jobReporterId);
-                    }
-                    boolean typeUpdated = false;
-                    if (!hypervisorType.getId().equals(knownHost.getTypeId())) {
-                        typeUpdated = true;
-                        knownHost.setType(hypervisorType);
-                    }
-
-                    GuestMigration guestMigration = new GuestMigration(consumerCurator)
-                        .buildMigrationManifest(incoming, knownHost);
-
-                    boolean factsUpdated = consumerResource.checkForFactsUpdate(knownHost, incoming);
-
-                    if (factsUpdated || guestMigration.isMigrationPending() || typeUpdated ||
-                        hypervisorIdUpdated) {
-                        knownHost.setLastCheckin(new Date());
-                        guestMigration.migrate(false);
-                        result.addUpdated(this.translator.translate(knownHost, HypervisorConsumerDTO.class));
-                    }
-                    else {
-                        result.addUnchanged(
-                            this.translator.translate(knownHost, HypervisorConsumerDTO.class));
-                    }
-                }
-                // update reporter id if it changed
-                if (jobReporterId != null && reportedOnConsumer != null &&
-                    reportedOnConsumer.getHypervisorId() != null &&
-                    (reportedOnConsumer.getHypervisorId().getReporterId() == null ||
-                        !jobReporterId.contentEquals(reportedOnConsumer.getHypervisorId().getReporterId()))) {
-                    reportedOnConsumer.getHypervisorId().setReporterId(jobReporterId);
-                }
-                else if (jobReporterId == null) {
-                    log.debug("hypervisor checkin reported asynchronously without reporter id " +
-                        "for hypervisor:{} of owner:{}", hypervisorId, ownerKey);
+                    updated.add(consumer);
                 }
             }
 
-            for (Consumer consumer : hypervisorKnownConsumersMap.getConsumers()) {
-                consumer = result.wasCreated(
-                    this.translator.translate(consumer, HypervisorConsumerDTO.class)) ?
-                    consumerCurator.create(consumer, false) :
-                    consumerCurator.update(consumer, false);
-            }
-
-            consumerCurator.flush();
+            doInBulk(created, consumers -> consumerCurator.saveAll(consumers, false, false));
+            doInBulk(updated, consumers -> consumerCurator.bulkUpdate(consumers, false));
 
             log.info("Summary for report from {} by principal {}\n {}", jobReporterId, principal, result);
             context.setResult(result);
@@ -367,16 +192,18 @@ public class HypervisorUpdateJob extends KingpinJob {
         }
     }
 
-    private void logReporterWarning(String jobReporterId, Consumer knownHost, String hypervisorId,
-        String ownerKey) {
-        if (jobReporterId != null && knownHost.getHypervisorId() != null &&
-            hypervisorId.equalsIgnoreCase(knownHost.getHypervisorId().getHypervisorId()) &&
-            knownHost.getHypervisorId().getReporterId() != null &&
-            !jobReporterId.equalsIgnoreCase(knownHost.getHypervisorId().getReporterId())) {
-            log.debug("Reporter changed for Hypervisor {} of Owner {} from {} to {}",
-                hypervisorId, ownerKey, knownHost.getHypervisorId().getReporterId(),
-                jobReporterId);
-        }
+    private HypervisorList parsedHypervisors(final JobDataMap map) throws IOException {
+        final byte[] data = (byte[]) map.get(DATA);
+        final String json = decompress(data);
+        return mapper.readValue(json, HypervisorList.class);
+    }
+
+    private void doInBulk(final List<Consumer> created,
+        final java.util.function.Consumer<Set<Consumer>> action) {
+        Lists.partition(created, BULK_SIZE)
+            .stream()
+            .map(HashSet::new)
+            .forEach(action);
     }
 
     /**
@@ -403,21 +230,19 @@ public class HypervisorUpdateJob extends KingpinJob {
 
         // Not sure if this is the best way to go:
         // Give each job a UUID to ensure that it is unique
-        JobDetail detail = newJob(HypervisorUpdateJob.class)
+        return newJob(HypervisorUpdateJob.class)
             .withIdentity(prefix + Util.generateUUID())
             .requestRecovery(true) // recover the job upon restarts
             .usingJobData(map)
             .storeDurably(true) // required if we have to postpone the job
             .build();
-
-        return detail;
     }
 
-    public static byte[] compress(String text) {
+    private static byte[] compress(String text) {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         try {
             OutputStream out = new DeflaterOutputStream(baos);
-            out.write(text.getBytes("UTF-8"));
+            out.write(text.getBytes(StandardCharsets.UTF_8));
             out.close();
         }
         catch (IOException e) {
@@ -426,7 +251,7 @@ public class HypervisorUpdateJob extends KingpinJob {
         return baos.toByteArray();
     }
 
-    public static String decompress(byte[] bytes) {
+    private static String decompress(byte[] bytes) {
         InputStream in = new InflaterInputStream(new ByteArrayInputStream(bytes));
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         try {
@@ -435,73 +260,11 @@ public class HypervisorUpdateJob extends KingpinJob {
             while ((len = in.read(buffer)) > 0) {
                 baos.write(buffer, 0, len);
             }
-            return new String(baos.toByteArray(), "UTF-8");
+            return new String(baos.toByteArray(), StandardCharsets.UTF_8);
         }
         catch (IOException e) {
             throw new AssertionError(e);
         }
-    }
-
-    /*
-     * Create a new hypervisor type consumer to represent the incoming hypervisorId
-     */
-    private Consumer createConsumerForHypervisorId(String incHypervisorId, String reporterId,
-        Owner owner, Principal principal, Consumer incoming) {
-        Consumer consumer = new Consumer();
-        if (incoming.getName() != null) {
-            consumer.setName(incoming.getName());
-        }
-        else {
-            consumer.setName(sanitizeHypervisorId(incHypervisorId));
-        }
-        consumer.setType(hypervisorType);
-        consumer.setFact("uname.machine", "x86_64");
-        consumer.setGuestIds(new ArrayList<>());
-        consumer.setLastCheckin(new Date());
-        consumer.setOwner(owner);
-        consumer.setAutoheal(true);
-        consumer.setCanActivate(subAdapter.canActivateSubscription(consumer));
-        if (owner.getDefaultServiceLevel() != null) {
-            consumer.setServiceLevel(owner.getDefaultServiceLevel());
-        }
-        else {
-            consumer.setServiceLevel("");
-        }
-        if (principal.getUsername() != null) {
-            consumer.setUsername(principal.getUsername());
-        }
-        consumer.setEntitlementCount(0L);
-        // TODO: Refactor this to not call resource methods directly
-        consumerResource.sanitizeConsumerFacts(consumer);
-
-
-        // Create HypervisorId
-        HypervisorId hypervisorId = new HypervisorId(consumer, owner, incHypervisorId);
-        hypervisorId.setReporterId(reporterId);
-        consumer.setHypervisorId(hypervisorId);
-
-        // TODO: Refactor this to not call resource methods directly
-        consumerResource.checkForFactsUpdate(consumer, incoming);
-
-        return consumer;
-    }
-
-    /*
-     * Make sure the HypervisorId is a valid consumer name.
-     */
-    private String sanitizeHypervisorId(String incHypervisorId) {
-        // Same validation as consumerResource.checkConsumerName
-        if (incHypervisorId.indexOf('#') == 0) {
-            log.debug("Hypervisor id cannot begin with # character");
-            incHypervisorId = incHypervisorId.substring(1);
-        }
-
-        int max = Consumer.MAX_LENGTH_OF_CONSUMER_NAME;
-        if (incHypervisorId.length() > max) {
-            log.debug("Hypervisor id too long, truncating");
-            incHypervisorId = incHypervisorId.substring(0, max);
-        }
-        return incHypervisorId;
     }
 
     /**
@@ -521,10 +284,4 @@ public class HypervisorUpdateJob extends KingpinJob {
         }
     }
 
-    public Consumer updateCheckinTime(Consumer consumer) {
-        Date now = new Date();
-        consumerCurator.updateLastCheckin(consumer, now);
-        consumer.setLastCheckin(now);
-        return consumer;
-    }
 }
